Дефинирана Node и WebWorker конфигурация
