+++ package.json
@@ -18,5 +18,9 @@
   "dependencies": {
     "dotenv": "^16.6.1",
     "jimp": "^0.22.12"
+  },
+  "devDependencies": {
+    "@types/jimp": "^0.2.1",
+    "@types/node": "^24.3.0"
   }
 }
